refactor: move joinXMPPChatRoom out of the SDK into AppServiceRoomService
The SDK's Application Service block is otherwise thin pass-throughs; this was
the only method with embedded provisioning logic. Extract it into a dedicated
AppServiceRoomService (federation-sdk, the layer that can see both the
appservice toolkit and the room/directory services) and leave a thin
pass-through on the SDK.
